ai_dedup_service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
datasketch
simhash
pdfminer.six
pypdf
python-dotenv
